analytics/requirements.txt
psycopg2-binary
python-dotenv
Faker
pytz